[#S11P21S004-440] [BE] feat: 실시간 피드백 관련 파일명 변경 및 coachIndex 자료형 변경 int -> Long 추가

+++ backend/src/main/java/com/pacemaker/domain/realtime/service/RealTimeService.java
@@ -14,7 +14,7 @@ public class RealTimeService {
 		this.gpuServerWebClient = gpuServerWebClient;
 	}
 
-	public Mono<byte[]> createRealTimeTts(String message, int coachIndex) {
+	public Mono<byte[]> createRealTimeTts(String message, Long coachIndex) {
 		return gpuServerWebClient.get()
 			.uri(uriBuilder -> uriBuilder
 				.path("/tts")
